+++ common/rfb/VNCSConnectionST.cxx
@@ -1706,7 +1706,7 @@ void VNCSConnectionST::sendNetworkStats() {
 
     fmt::format_to(std::back_inserter(buf), "[{}, {}, {}]", jitter, congestion.getPingTime(), congestion.getBandwidth());
     vlog.info("Sending diagnostic network stats:\n%s\n", buf.data());
-    writer()->writeStats(msgTypeNetworkStats, buf.data(), static_cast<int>(buf.size()));
+    writer()->writeStats(msgTypeNetworkStats, buf.data(), buf.size());
 }
 
 void VNCSConnectionST::sendSystemStats() {
